+++ CHANGES.md
@@ -11,6 +11,10 @@
 
    - <https://github.com/georust/gdal/pull/494>
 
+- Implemented `Feature::set_field_null`
+
+  - <https://github.com/georust/gdal/pull/501>
+
 - **Breaking**: Changed a number of APIs using `isize` when `usize` is semantically more appropriate: `Driver::create.*`, `Rasterband::overview`, `Dataset::{layer|into_layer|layer_count}`.  
 
   - <https://github.com/georust/gdal/pull/497>